I moved my new formAPI utility object into a new .js module and exported/imported accordingly
